+++ src/views/reports/ListReports.vue
@@ -202,7 +202,7 @@ const updateSortByParams = (params: RouteParamsGeneric) => {
   if (params.sortBy) {
     sortByMap.value.set(
       params.sortBy as keyof Moneyflow,
-      params.sortDirection === "desc",
+      params.sortDirection === "asc",
     );
   } else {
     sortByMap.value.set("bookingDate", false);
